Refactor codefix: reduce code duplication
