+++ lib/main/atom/codefix/index.ts
@@ -1,0 +1,3 @@
+export {CodeActionsProvider} from "./codeActionsProvider"
+export {IntentionsProvider} from "./intentionsProvider"
+export {CodefixProvider} from "./codefixProvider"
